gradio_app/pyproject.toml
[tool.poetry]
name = "gradio-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
gradio = "^5.8.0"
python-dotenv = "^1.0.1"
authlib = "^1.3.2"
itsdangerous = "^2.2.0"
ruff = "^0.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
